Move: ServiceModule로 Service provider 이동 (#9)
