+++ app/src/main/java/com/infinity/omos/di/NetworkModule.kt
@@ -2,10 +2,6 @@ package com.infinity.omos.di
 
 import com.infinity.omos.BuildConfig
 import com.infinity.omos.api.AuthService
-import com.infinity.omos.api.BlockService
-import com.infinity.omos.api.FakeUserService
-import com.infinity.omos.api.FollowService
-import com.infinity.omos.api.UserService
 import com.infinity.omos.utils.DataStoreManager
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import dagger.Module
@@ -15,19 +11,14 @@ import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import okhttp3.Authenticator
 import okhttp3.Interceptor
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
-import okhttp3.ResponseBody.Companion.toResponseBody
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Converter
 import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Named
 import javax.inject.Provider
 import javax.inject.Singleton
 
@@ -117,45 +108,5 @@ object NetworkModule {
             .client(okHttpClient)
             .addConverterFactory(converterFactory)
             .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideUserService(
-        retrofit: Retrofit
-    ): UserService {
-        return retrofit.create(UserService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideAuthService(
-        retrofit: Retrofit
-    ): AuthService {
-        return retrofit.create(AuthService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideBlockedService(
-        retrofit: Retrofit
-    ): BlockService {
-        return retrofit.create(BlockService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideFollowService(
-        retrofit: Retrofit
-    ): FollowService {
-        return retrofit.create(FollowService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideFakeUserService(
-        retrofit: Retrofit
-    ): FakeUserService {
-        return retrofit.create(FakeUserService::class.java)
     }
 }
